+++ src/option.c
@@ -8,7 +8,7 @@
  */
 
 
-#include "coap_config.h"
+#include "coap_internal.h"
 
 #if defined(HAVE_ASSERT_H) && !defined(assert)
 # include <assert.h>
@@ -16,13 +16,6 @@
 
 #include <stdio.h>
 #include <string.h>
-
-#include "libcoap.h"
-#include "option.h"
-#include "encode.h"                /* for coap_fls() */
-#include "coap_debug.h"
-#include "mem.h"
-#include "utlist.h"
 
 #define ADVANCE_OPT(o,e,step) if ((e) < step) {           \
     coap_log(LOG_DEBUG, "cannot advance opt past end\n"); \
