+++ src/subscribe.c
@@ -1,20 +1,17 @@
 /* subscribe.c -- subscription handling for CoAP
  *                see draft-ietf-coap-observe-16
  *
- * Copyright (C) 2010--2013,2015 Olaf Bergmann <bergmann@tzi.org>
+ * Copyright (C) 2010--2019 Olaf Bergmann <bergmann@tzi.org>
  *
  * This file is part of the CoAP library libcoap. Please see
  * README for terms of use.
  */
 
-#include "coap_config.h"
-#include "coap.h"
+#include "coap_internal.h"
 
 #if defined(HAVE_ASSERT_H) && !defined(assert)
 # include <assert.h>
 #endif
-
-#include "subscribe.h"
 
 void
 coap_subscription_init(coap_subscription_t *s) {
