+++ win32/libcoap.vcxproj.filters
@@ -24,13 +24,34 @@
     <ClCompile Include="..\src\block.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\coap_debug.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\src\coap_event.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\src\coap_hashkey.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\src\coap_gnutls.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\coap_io.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\src\coap_notls.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\src\coap_openssl.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\src\coap_session.c">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\src\coap_time.c">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\coap_debug.c">
+    <ClCompile Include="..\src\coap_tinydtls.c">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\src\encode.c">
@@ -60,24 +81,6 @@
     <ClCompile Include="..\src\uri.c">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\coap_event.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="..\src\coap_notls.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="..\src\coap_tinydtls.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="..\src\coap_openssl.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="..\src\coap_session.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="..\src\coap_hashkey.c">
-      <Filter>Source Files</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\coap_config.h">
@@ -98,6 +101,9 @@
     <ClInclude Include="..\include\coap2\coap.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\include\coap2\coap_debug.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="..\include\coap2\coap_dtls.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -107,13 +113,19 @@
     <ClInclude Include="..\include\coap2\coap_io.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\include\coap2\coap_mutex.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="..\include\coap2\coap_session.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\include\coap2\coap_time.h">
+    <ClInclude Include="..\include\coap2\coap_session_internal.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\include\coap2\coap_debug.h">
+    <ClInclude Include="..\include\coap2\coap_subscribe_internal.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\include\coap2\coap_time.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\include\coap2\encode.h">
